Correções nos nomes das classes e documentação do código desenvolvido até à data
