+++ .vs/VSWorkspaceState.json
@@ -1,6 +1,8 @@
 {
   "ExpandedNodes": [
-    ""
+    "",
+    "\\bin"
   ],
+  "SelectedNode": "\\Clientes.cs",
   "PreviewInSolutionExplorer": false
 }
